requirements.txt
# ── Serveur web ────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0

# ── Deep Learning — modèle EfficientNet+mBERT ─────────
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.0

# ── Transformers — mBERT tokenizer + modèle ───────────
transformers>=4.36.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# ── OCR bilingue (français + arabe) ───────────────────
easyocr>=1.7.0
opencv-python-headless>=4.8.0

# ── Traitement d'images ────────────────────────────────
Pillow>=10.0.0

# ── PDF → image ────────────────────────────────────────
pdf2image>=1.16.0
pymupdf>=1.23.0

# ── Base de données ────────────────────────────────────
pymongo>=4.6.0

# ── Appels HTTP (Ollama / Gemma2 / Qwen) ──────────────
requests>=2.31.0

# ── Utilitaires ────────────────────────────────────────
numpy>=1.24.0
scikit-learn>=1.3.0